refactor(era5): move spi processing out of main era5 module
